Remove RequestOptions
