+++ sdk/core/Azure.Core/src/ChainingClassifier.cs
@@ -13,7 +13,7 @@ namespace Azure.Core
     /// characteristics.
     /// The classifier chain is a series of <see cref="ResponseClassificationHandler"/> classifiers
     /// followed by the "end-of-chain" <see cref="ResponseClassifier"/>.  The handlers are
-    /// added to the chain via <see cref="RequestContext"/> or <see cref="RequestOptions"/>,
+    /// added to the chain via <see cref="RequestContext"/>,
     /// and all of them are applied starting with the most recently added handler and
     /// iterating over the list to the least-recently added handler, then applying status code
     /// classification, and finally by applying the "end-of-chain" classifier.
